requirements.txt
tensorflow-gpu
pycocotools
scikit-image